backend/requirements.txt
fastapi
uvicorn
numpy
nibabel
Pillow
torch
torchvision
torchaudio
nnunetv2
ultralytics
pydantic
radiomics
scikit-learn
pandas
SimpleITK
joblib
